backend/requirements.txt
# Web framework
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.0

# Database
psycopg2-binary==2.9.9

# AI/ML
tensorflow==2.15.0
numpy==1.26.4

# Utilities
python-dotenv==1.0.1
bcrypt==4.1.2
firebase-admin==6.4.0
pandas==2.2.1
gunicorn==21.2.0
pytest==8.0.2
black==24.2.0
flake8==7.0.0